Java design patterns were studied and software principles were learned.
Java design patterns were studied and software principles were learned.

SOLID software principles were learned. I learned creational, structural and behavioral design patterns.
The aim of this study is to ensure that the software we have developed can easily adapt to the requirements in the future, that we can easily add new features without the need for a change in the code, that it ensures minimal changes to the code despite new requirements, and that it also minimizes the loss of time caused by problems such as constant correction or even rewriting of the code.